Allow updating text after checking entries
A function will check the values of currently updated and the other field
entries. If any of these is empty, or if they are not equal, a non-match text
is kept. Otherwise, a match text is kept. Event listeners check for
changes to values of password or confirm password fields. If there
is a change, the update text function is called to check for a match